App completed

+++ lib/controller/cart_controller.dart
@@ -21,13 +21,22 @@ class CartController extends GetxController {
   var vendors = [];
 
   var placingOrder = false.obs;
-
   calculate(data) {
-    totalP.value = 0;
+    totalP.value = 0.0.toInt();
     for (var i = 0; i < data.length; i++) {
-      totalP.value = totalP.value + int.parse(data[i]['tprice'].toString());
+      totalP.value =
+          (totalP.value.toDouble() + double.parse(data[i]['tprice'].toString()))
+              .toInt();
     }
   }
+
+  // calculate(data) {
+  //   totalP.value = 0;
+  //   for (var i = 0; i < data.length; i++) {
+  //     totalP.value =
+  //         totalP.value + int.parse(data[i]['tprice'].toString());
+  //   }
+  // }
 
   changePaymentIndex(index) {
     paymentIndex.value = index;
